+++ GOL.contract/gameEngine.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+
+namespace GOL.Contract
+{
+    public interface IGameEngine
+    {
+        List<Cell> NextGeneration(List<Cell> grid, int width, int height);
+    }
+
+}
