+++ GOL/Gameoflife.cs
@@ -1,55 +1,59 @@
+
+
+using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading;
+using GOL.Business;
 using GOL.Contract;
 
 namespace GOL
 {
-    /// <summary>
-    /// Kernlogik des Game of Life implementiert das IGameEngine-Contract.
-    /// </summary>
-    public class GameEngine : IGameEngine
+       
+
+     internal class Programm
     {
-        public List<Cell> NextGeneration(List<Cell> grid, int width, int height)
+        private static void Render(List<Cell> grid, int width, int height)
         {
-            return Enumerable.Range(0, width)
-                             .SelectMany(x => Enumerable.Range(0, height)
-                                                        .Select(y =>
-                                                        {
-                                                            var current = grid.FirstOrDefault(c => c.X == x && c.Y == y)
-                                                                          ?? new Cell { X = x, Y = y, IsAlive = false };
-                                                            int aliveNeighbors = CountNeighbors(grid, current, width, height);
-                                                            bool nextAlive = aliveNeighbors == 3
-                                                                             || (aliveNeighbors == 2 && current.IsAlive);
-                                                            return new { current, nextAlive };
-                                                        }))
-                             .Where(t => t.nextAlive)
-                             .Select(t => new Cell
-                             {
-                                 X = t.current.X,
-                                 Y = t.current.Y,
-                                 IsAlive = true
-                             })
-                             .ToList();
-        }
+            Console.SetCursorPosition(0, 0);
+            Console.Clear();
+            Console.WriteLine("\x1b[3J");
 
-        public int CountNeighbors(List<Cell> grid, Cell cell, int width, int height)
-        {
-            return Enumerable.Range(-1, 3)
-                             .SelectMany(dx => Enumerable.Range(-1, 3), (dx, dy) => (dx, dy))
-                             .Where(d => d.dx != 0 || d.dy != 0)
-                             .Select(d =>
-                             {
-                                 int nx = (cell.X + d.dx + width) % width;
-                                 int ny = (cell.Y + d.dy + height) % height;
-                                 return grid.FirstOrDefault(n => n.X == nx && n.Y == ny);
-                             })
-                             .Count(n => n != null && n.IsAlive);
+            for (int y = 0; y < height; y++)
+            {
+                for (int x = 0; x < width; x++)
+                {
+                    bool alive = grid.Any(c => c.X == x && c.Y == y && c.IsAlive);
+                    Console.Write(alive ? "O " : ". ");
+                }
+                Console.WriteLine();
+            }
         }
-
-        public bool ValidateExistence(List<Cell> grid, Cell cell, int width, int height)
+        public static void Start(int width, int height)
         {
-            int neighbors = CountNeighbors(grid, cell, width, height);
-            return neighbors == 3 || (neighbors == 2 && cell.IsAlive);
+            IGameEngine engine = new GameEngine();
+
+            // Beispielgrid mit definierten lebenden Zellen
+            var grid = new List<Cell>
+            {
+                new Cell { X = 1, Y = 1, IsAlive = true },
+                new Cell { X = 2, Y = 1, IsAlive = true },
+                new Cell { X = 4, Y = 1, IsAlive = true },
+                new Cell { X = 5, Y = 1, IsAlive = true },
+                new Cell { X = 6, Y = 1, IsAlive = true },
+                new Cell { X = 7, Y = 1, IsAlive = true },
+                new Cell { X = 8, Y = 1, IsAlive = true },
+                new Cell { X = 3, Y = 1, IsAlive = true }
+            };
+
+            // Hauptschleife zur Simulation
+            while (true)
+            {
+                Render(grid, width, height);
+                grid = engine.NextGeneration(grid, width, height);
+                Thread.Sleep(500); // 500ms Pause für Geschwindigkeit
+            }
         }
     }
+    
 }
